third_party/blink/web_tests/external/wpt/selection/selection-direction-on-triple-click.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] direction returns "none" when there is a triple click selection(directionless)
  assert_equals: expected Element node <div id="container">hello, world</div> but got Text node "hello, world"
Harness: the test ran to completion.
